Breaking out mage files
