+++ .github/workflows/ci-code-approval.yml
@@ -38,11 +38,6 @@ jobs:
       - name: Checkout repository
         uses: actions/checkout@v4
 
-      - name: "Configure git for private modules"
-        env:
-          GITHUB_API_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-        run: git config --global url."https://x:${GITHUB_API_TOKEN}@github.com".insteadOf "https://github.com"
-
       - name: golangci-lint
         uses: golangci/golangci-lint-action@v6
         with:
@@ -134,7 +129,7 @@ jobs:
         uses: magefile/mage-action@v3
         with:
           version: latest
-          args: apilint
+          args: linting:apis
 
       - name: Upload the report
         # Run this step on all PR's even if the linting fails
